renaming profile as an example
