+++ shared/nav.desktop.js
@@ -11,7 +11,7 @@ import Chat from './chat'
 import People from './people'
 import Devices from './devices'
 import NoTab from './no-tab'
-import Profile from './profile'
+import Profile from './profile/container'
 import Search from './search'
 import Settings from './settings'
 import Login from './login'
