+++ shared/profile/dumb.js
@@ -2,7 +2,7 @@
 import ConfirmOrPending from './confirm-or-pending'
 import EditAvatar from './edit-avatar'
 import PostProof from './post-proof'
-import Profile from './render'
+import Profile from './index'
 import ProveEnterUsername from './prove-enter-username'
 import ProveWebsiteChoice from './prove-website-choice'
 import Revoke from './revoke'
@@ -15,7 +15,7 @@ import {normal, checking, revoked, error, metaNone, metaNew, metaDeleted, metaPe
 
 import type {DumbComponentMap} from '../constants/types/more'
 import type {Proof} from '../constants/tracker'
-import type {Props as RenderProps} from './render'
+import type {Props as RenderProps} from './index'
 import type {UserInfo} from '../common-adapters/user-bio'
 
 export const proofsDefault: Array<Proof> = [
