+++ shared/profile/revoke/index.native.js
@@ -1,9 +1,9 @@
 /* @flow */
 
 import React from 'react'
-import {Text} from '../common-adapters'
+import {Text} from '../../common-adapters'
 
-import type {Props} from './confirm-or-pending'
+import type {Props} from './index'
 
 const Render = ({platform, platformHandle, onCancel, onRevoke}: Props) => {
   return <Text type='Body'>TODO: Revoke</Text>
