+++ shared/search/user-pane/non-user.render.desktop.js
@@ -5,7 +5,7 @@ import {Avatar, Box, Icon, Text} from '../../common-adapters'
 import {globalColors, globalStyles, globalMargins} from '../../styles'
 import {capitalize} from 'lodash'
 import {platformToLogo24} from '../../constants/search'
-import {AVATAR_SIZE, HEADER_TOP_SPACE, HEADER_SIZE} from '../../profile/render.desktop'
+import {AVATAR_SIZE, HEADER_TOP_SPACE, HEADER_SIZE} from '../../profile/index.desktop'
 import type {Props} from './non-user.render'
 
 import electron from 'electron'
